% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{AgreementBudget-class}
\alias{AgreementBudget-class}
\title{Class AgreementBudget}
\description{
An S4 class for information about sources of agreement and disagreement
between three categorical raster maps.
}
\section{Slots}{

\describe{
\item{\code{agreement}}{data.frame containing sources of agreement and disagreement.
Rows represent different resolutions and columns represent the different
types of agreement and disagreement}

\item{\code{factors}}{numeric vector of aggregation factors}

\item{\code{type}}{character. Either 'overall' (considering all possible transitions),
'category' (all transitions from one category) or 'transition' (specific
transition)}

\item{\code{categories}}{numeric vector of land use categories in the three
maps considered}

\item{\code{labels}}{character vector with labels corresponding to \code{categories}}

\item{\code{from,to}}{numeric vectors that together define the land use transitions
  for which agreement was calculated}
}}

