% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{AgreementBudget.plot}
\alias{AgreementBudget.plot}
\alias{AgreementBudget.plot,AgreementBudget-method}
\title{Plot method for AgreementBudget objects}
\usage{
AgreementBudget.plot(x, ...)

\S4method{AgreementBudget.plot}{AgreementBudget}(x,
  col = RColorBrewer::brewer.pal(5, "Set2"), key, scales, xlab, ylab, ...)
}
\arguments{
\item{x}{object of class \code{AgreementBudget}}

\item{col}{character specifying the plotting colour. Default is to use the 'Set2'
palette from \code{RColorBrewer}}

\item{key}{list. See \code{lattice::\link[lattice]{xyplot}}}

\item{scales}{list. See \code{lattice::\link[lattice]{xyplot}}}

\item{xlab}{character or expression. See \code{lattice::\link[lattice]{xyplot}}}

\item{ylab}{character or expression. See \code{lattice::\link[lattice]{xyplot}}}

\item{\dots}{additional arguments to \code{lattice::\link[lattice]{xyplot}}}
}
\value{
A trellis object.
}
\description{
Plot an \code{\link{AgreementBudget}} object.
}
\details{
The plot layout is based on work presented in Pontius et al. (2011)
}
\references{
Pontius Jr, R.G., Peethambaram, S., Castella, J.C. (2011).
Comparison of three maps at multiple resolutions: a case study of land change
simulation in Cho Don District, Vietnam. Annals of the Association of American
Geographers 101(1): 45-62.
}
\seealso{
\code{\link{AgreementBudget}}, \code{lattice::\link[lattice]{xyplot}}
}

