% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AgreementBudget}
\alias{AgreementBudget}
\title{Create an AgreementBudget object}
\usage{
AgreementBudget(x, from, to, ...)
}
\arguments{
\item{x}{a \code{ThreeMapComparison} object}

\item{from}{numeric (optional). A single value representing a land use
category. Results will be restricted to transitions from this category}

\item{to}{numeric (optional). Similar to \code{from}. If provided with a valid
\code{from} argument the result will be restricted to the transition
defined by these two arguments (i.e. \code{from} -> \code{to})}

\item{...}{additional arguments (none)}
}
\value{
an \code{AgreementBudget} object
}
\description{
This function quantifies sources of agreement and disagreement between a
reference map for time 1, a reference map for time 2 and a simulated map for
time 2 to provide meaningful information about the performance of land use
change simulations.
}
\details{
The types of agreement and disagreement considered are as follows:

\enumerate{
  \item Persistence simulated correctly (agreement)
  \item Persistence simulated as change (disagreement)
  \item Change simulated incorrectly (disagreement)
  \item Change simulated correctly (agreement)
  \item Change simulated as persistence (disagreement)
}
}
\author{
Simon Moulds
}
\seealso{
\code{link{ThreeMapComparison}}, \code{link{FigureOfMerit}}
}

