% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allow}
\alias{allow}
\title{Implement decision rules for land use change}
\usage{
allow(x, hist, categories, cd, rules, ...)
}
\arguments{
\item{x}{numeric vector containing the land use pattern for the current
timestep}

\item{hist}{numeric vector containing land use history (values represent the
number of years the cell has contained the current land use category)}

\item{categories}{numeric vector containing land use categories in the study
region}

\item{cd}{numeric vector indicating the direction of change for each
land use category. A value of 1 means demand is increasing (i.e. the number
of cells belonging to the category must increase), -1 means decreasing
demand and 0 means demand is static}

\item{rules}{matrix. See details}

\item{...}{additional arguments (none)}
}
\value{
a matrix with values of 1 (change allowed) or NA (change not allowed)
}
\description{
Identify legitimate transitions based on land use history and specific
transition rules.
}
\details{
Decision rules are based on those described by Verburg et al. (2002). The
\code{rules} input argument is a square matrix with dimensions equal to the
number of land use categories in the study region where rows represent the
current land use and columns represent future transitions. The value of each
element should represent a rule from the following list:

\enumerate{
  \item (rule == 0 | rule == 1): this rule concerns specific land use
    transitions that are allowed (1) or not (0)
  \item (rule == -1): this rule prevents transitions unless demand
    for the present land use category is decreasing
  \item (rule == -2): this rule prevents transitions to land use
    categories with decreasing or static demand. Note that by combining
    rule 3 and rule 2 it is possible to prevent simultaneous expansion and
    contraction
  \item (rule > 100 & rule < 1000): this rule imposes a time limit (rule-100)
    on land use transitions, after which land use change is not allowed. Time
    is taken from \code{hist}
  \item (rule > 1000): this rule imposes a minimum period of time (rule-1000)
    before land use is allowed to change
}

\code{allow} should be called from methods in \code{\link{allocate}}. The
output is a matrix with the same dimensions as the matrix used internally by
allocation functions to store land use suitability. Thus, by multiplying the
two matrices together, disallowed transitions can be removed from the
allocation procedure.
}
\examples{
obs <- ObservedMaps(x=pie,
                    pattern="lu",
                    categories=c(1,2,3),
                    labels=c("forest","built","other"),
                    t=c(0,6,14))

# create arbitrary land use history raster
hist <- raster(obs@maps[[1]])
vals <- sample(1:10, ncell(hist))
hist <- setValues(hist, vals[which(!is.na(getValues(obs@maps[[1]])))]

# calculate demand and get change direction for first timestep
dmd <- approxExtrapDemand(obs=obs, tout=0:14)
cd <- dmd[2,] - dmd[1,]

# create rules matrix, only allowing forest to change if the cell has
# belonged to forest for more than 8 years
rules <- matrix(data=c(1,1008,1008,
                       1,1,1,
                       1,1,1), nrow=3, ncol=3, byrow=TRUE)

allow <- allow(x=obs@maps[[1]],
               hist=hist,
               categories=obs@categories,
               cd=cd,
               rules=rules)

# NB output is only useful when used within an allocation routine
}
\author{
Simon Moulds
}
\references{
Verburg, P.H., Soepboer, W., Veldkamp, A., Limpiada, R., Espaldon,
V., Mastura, S.S. (2002). Modeling the spatial dynamics of regional land use:
the CLUE-S model. Environmental management, 30(3):391-405.
}

