% Generated by roxygen2 (4.0.2): do not edit by hand
\name{approxExtrapDemand}
\alias{approxExtrapDemand}
\title{Extrapolate land use area}
\usage{
approxExtrapDemand(obs, tout, ...)
}
\arguments{
\item{obs}{an \code{ObservedMaps} object containing at least two maps}

\item{tout}{numeric vector specifying the timesteps where interpolation is to
take place. Comparable to the \code{xout} argument of
\code{Hmisc::\link[Hmisc]{approxExtrap}}}

\item{...}{additional arguments to \code{Hmisc::\link[Hmisc]{approxExtrap}}}
}
\value{
a matrix
}
\description{
Extrapolate land use area from two observed land maps to provide
a valid (although not necessarily realistic) demand scenario.
}
\examples{
obs <- ObservedMaps(x=pie,
                    pattern="lu",
                    categories=c(1,2,3),
                    labels=c("forest","built","other"),
                    t=c(0,6,14))
dmd1 <- approxExtrapDemand(obs=obs, tout=c(0:14))
}
\author{
Simon Moulds
}

