% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.Predictors}
\alias{as.data.frame.Predictors}
\title{Coerce \code{Predictors} object to data.frame}
\usage{
as.data.frame.Predictors(x, row.names = NULL, optional = FALSE, cells,
  timestep = 0, ...)
}
\arguments{
\item{x}{an object of class \code{\link{Predictors}}}

\item{row.names}{NULL or a character vector giving the row.names for the
data.frame. Missing values are not allowed}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional}

\item{cells}{index of cells to be extracted}

\item{timestep}{numeric indicating the timestep under consideration. Only
relevant if x@maps contains dynamic predictor variables}

\item{...}{additional arguments (none)}
}
\value{
data.frame
}
\description{
This function extracts data from all rasters in a \code{\link{Predictors}}
object for a specified timestep (if dynamic or \code{\link{PredictorCall}}
objects are present).
}
\author{
Simon Moulds
}
\seealso{
\code{\link{Predictors}}, \code{\link{predictorMaps}},
\code{\link{PredictorCall}}, \code{\link{partition}}
}

