% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoConvert}
\alias{autoConvert}
\title{Automatic land use transitions}
\usage{
autoConvert(x, mask, prob, categories, ...)
}
\arguments{
\item{x}{numeric vector containing the land use pattern for the current
timestep}

\item{mask}{numeric vector containing binary values where 0 indicates cells
that are not allowed to change. These are automatically assigned the same
land use category for every timestep}

\item{prob}{matrix. TODO}

\item{categories}{numeric vector containing land use categories}

\item{...}{additional arguments (none)}
}
\value{
RasterLayer or numeric vector
}
\description{
Perform automatic land use transitions in cells that have been excluded from
the main allocation routine. This might happen if cells lie within masked
areas and cannot change land use type (in which case \code{autoConvert} sets
land use to the land use for the previous timestep), or because they must
change according to decision rules (in which case the cell is allocated to the
land use with the highest suitability value).
}

