% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{CategoryLabel-class}
\alias{CategoryLabel-class}
\title{Virtual class Model}
\description{
A virtual S4 class to represent land use change models.
}
\section{Slots}{

\describe{
\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

