% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FigureOfMerit-class}
\alias{FigureOfMerit-class}
\title{Class FigureOfMerit}
\description{
An S4 class for different figure of merit scores.
}
\section{Slots}{

\describe{
\item{\code{overall}}{list containing the overall figure of merit score for each
resolution}

\item{\code{category}}{list of numeric vectors containing category specific scores}

\item{\code{transition}}{list of matrices containing transition specific scores}

\item{\code{factors}}{numeric vector of aggregation factors}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}
\author{
Simon Moulds
}

