% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ObservedMaps-class}
\alias{ObservedMaps-class}
\title{Class ObservedMaps}
\description{
An S4 class for observed land use maps.
}
\section{Slots}{

\describe{
\item{\code{maps}}{RasterStack containing observed land use maps}

\item{\code{t}}{numeric vector with timesteps corresponding to each observed map}

\item{\code{total}}{matrix with number of cells belonging to each land use category in
the observed maps. Rows represent different maps and columns represent land
land use categories}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}
\author{
Simon Moulds
}

