% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{PredictorCall-class}
\alias{PredictorCall-class}
\title{Class PredictorCall}
\description{
An S4 class for predictor variables derived from the observed pattern of land
use that should be updated during a simulation when a new land use map is
available.
}
\section{Slots}{

\describe{
\item{\code{map}}{RasterLayer showing the predictor variable}

\item{\code{call}}{function call used to calculate predictor variable}

\item{\code{update.arg}}{the argument (character) in \code{call} that should be
updated when a new land use map is available}

\item{\code{name}}{the variable name (character)}
}}
\author{
Simon Moulds
}

