% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Predictors-class}
\alias{Predictors-class}
\title{Class Predictors}
\description{
An S4 class for predictor variables.
}
\section{Slots}{

\describe{
\item{\code{maps}}{list of RasterStacks. The length of the list corresponds to the
number of predictor variables and the number of layers in each RasterStack
represents time}

\item{\code{calls}}{list of PredictorCall objects}

\item{\code{map.names}}{character vector of the name of each variable in \code{maps}}

\item{\code{call.names}}{character vector of the name of each variable in \code{calls}}

\item{\code{dynamic}}{logical indicating whether dynamic variables are present}
}}
\author{
Simon Moulds
}

