% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{compareAUC}
\alias{compareAUC}
\alias{compareAUC,Prediction-method}
\alias{compareAUC,list-method}
\title{Compare the area under the ROC curve (AUC) for different predictive models}
\usage{
compareAUC(pred, ...)

\S4method{compareAUC}{Prediction}(pred, digits = 4, ...)

\S4method{compareAUC}{list}(pred, digits = 4, ...)
}
\arguments{
\item{pred}{a \code{Prediction} object or a list of these}

\item{digits}{numeric indicating the number of digits to be displayed after
the decimal point for AUC values}

\item{\dots}{additional arguments (none)}
}
\value{
data.frame containing AUC values
}
\description{
Estimate the AUC for each \code{ROCR::\link[ROCR]{prediction}} object in a
\code{\link{Prediction}} object.
}
\details{
The user can compare the performance of different statistical models by
providing a list of \code{Prediction} objects. Note that
\code{compareAUC} should be used in conjunction with other comparison methods
because the AUC does not contain as much information as, for instance, the ROC
curve itself
}
\references{
Sing, T., Sander, O., Beerenwinkel, N., Lengauer, T. (2005).
ROCR: visualizing classifier performance in R. Bioinformatics
21(20):3940-3941.
}
\seealso{
\code{\link{Prediction}}, \code{ROCR::\link[ROCR]{performance}}
}

