% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{crossTabulate}
\alias{crossTabulate}
\alias{crossTabulate,ObservedMaps,ANY-method}
\alias{crossTabulate,RasterLayer,RasterLayer-method}
\title{Cross tabulate land use transitions}
\usage{
crossTabulate(x, y, ...)

\S4method{crossTabulate}{RasterLayer,RasterLayer}(x, y, categories,
  labels = as.character(categories), ...)

\S4method{crossTabulate}{ObservedMaps,ANY}(x, y, index, ...)
}
\arguments{
\item{x}{RasterLayer representing land use map from an earlier timestep or an
ObservedMaps object containing at least two land use maps for different
points in time}

\item{y}{RasterLayer representing land use map from a later timestep. Not used
if \code{x} is an ObservedMaps object}

\item{...}{additional arguments to \code{raster::\link[raster]{crosstab}}}

\item{categories}{numeric vector containing land use categories to consider.
Not used if \code{x} is an ObservedMaps object}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}. Not used if \code{x} is an ObservedMaps object}

\item{index}{numeric vector with index of land use maps from ObservedMaps
to use in analysis}
}
\value{
a data.frame
}
\description{
Cross tabulate land use transitions using
\code{raster::\link[raster]{crosstab}}. This should be the basis for further
research into the processes driving the most important transitions in the
study region (Pontius et al., 2004).
}
\examples{
# RasterLayer input
crossTabulate(x=pie$lu_pie_1985,
              y=pie$lu_pie_1999,
              categories=c(1,2,3),
              labels=c("forest","built","other"))

# ObservedMaps input
obs <- ObservedMaps(x=pie,
                    pattern="lu",
                    categories=c(1,2,3),
                    labels=c("forest","built","other"),
                    t=c(0,6,14))

crossTabulate(x=obs, index=c(1,3))
}
\author{
Simon Moulds
}
\references{
Pontius Jr, R.G., Shusas, E., McEachern, M. (2004). Detecting
important categorical land changes while accounting for persistence.
Agriculture, Ecosystems & Environment 101(2):251-268.
}
\seealso{
\code{\link{ObservedMaps}}, \code{raster::\link[raster]{crosstab}}
}

