% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ExpVarMaps-class}
\alias{ExpVarMaps-class}
\title{Class ExpVarMaps}
\description{
An S4 class for explanatory variables.
}
\section{Slots}{

\describe{
\item{\code{maps}}{list of RasterStack objects. The length of the list corresponds to
the number of explanatory variables and the number of layers in each
RasterStack represents time}

\item{\code{varnames}}{character vector with the name of each variable in \code{maps}}

\item{\code{dynamic}}{logical indicating whether dynamic variables are present}
}}

