% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ExpVarMaps}
\alias{ExpVarMaps}
\alias{ExpVarMaps,RasterStack,character-method}
\alias{ExpVarMaps,RasterStack-method}
\alias{ExpVarMaps,character,character-method}
\alias{ExpVarMaps,character-method}
\alias{ExpVarMaps,list,character-method}
\alias{ExpVarMaps,list-method}
\alias{ExpVarMaps,missing,character-method}
\alias{ExpVarMaps,missing-method}
\title{Create an ExpVarMaps object}
\usage{
ExpVarMaps(x, ...)

\S4method{ExpVarMaps}{missing}(x, pattern = NULL, ...)

\S4method{ExpVarMaps}{character}(x, pattern = NULL, ...)

\S4method{ExpVarMaps}{RasterStack}(x, pattern = NULL, ...)

\S4method{ExpVarMaps}{list}(x, pattern = NULL, ...)
}
\arguments{
\item{x}{path (character) to directory containing observed land use maps,
a Raster* object or a list of Raster* objects}

\item{pattern}{regular expression (character). Only filenames (if \code{x} is
a path) or Raster* objects (if \code{x} is a list) matching the regular
expression will be returned. See \cr
\code{raster::\link[raster]{raster}} for more information about supported
filetypes}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
An ExpVarMaps object.
}
\description{
Methods to load maps of explanatory variables, which may be created from file,
an existing Raster* object or a list of Raster* objects.
}
\details{
Explanatory variables should follow a naming convention to identify them as
static (one map provided for the study period) or dynamic (one map provided
for each year of the study period). The name should consist of two (static)
or three (dynamic) parts: firstly, the prefix should differentiate explanatory
variables from other maps in the directory, list or RasterStack. This should
be followed by a unique number to differentiate the explanatory variables
(note that the order of variables in the ExpVarMaps object is determined by
this value) If the variable is dynamic this number should be followed by a
second number representing the timestep to which the map applies. Dynamic
variables should include a map for time 0 (corresponding to the initial
observed map) and every subsequent timestep in the simulation. The different
parts should be separated by a period or underscore.

Maps of different explanatory variables should have the same coordinate
reference system but do not have to have the same extent and resolution as
long as the minimum extent is that of the study region defined by an
\code{ObsLulcMaps} object. However, maps for different timesteps of the same
dynamic variable should have the same extent and resolution because these are
stored as RasterStack objects.
}
\examples{
## Plum Island Ecosystems
ef <- ExpVarMaps(x=pie, pattern="ef")

## Sibuyan
ef <- ExpVarMaps(x=sibuyan$maps, pattern="ef")
}
\seealso{
\code{raster::\link[raster]{stack}}
}

