% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Extract by index}
\alias{Extract by index}
\alias{[[,CategoryLabel,ANY,ANY-method}
\alias{[[,ExpVarMaps,ANY,ANY-method}
\title{Extract by index}
\usage{
\S4method{[[}{ExpVarMaps,ANY,ANY}(x, i, j, ...)

\S4method{[[}{CategoryLabel,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{an object of class ExpVarMaps or any object inheriting from the
virtual class CategoryLabel}

\item{i}{layer number (if 'x' inherits from a RasterStack) or list index (if
'x' stores data as a list)}

\item{j}{numeric (not used)}

\item{...}{additional arguments (none)}
}
\description{
\code{object[[i]]} can be used to extract individual objects from container
classes such as \code{ExpVarMaps}, \code{PredModels}, \code{Prediction} and
\code{Performance}.
}
\examples{
## Plum Island Ecosystems

## Load observed land use maps
obs <- ObsLulcMaps(x=pie,
                   pattern="lu",
                   categories=c(1,2,3),
                   labels=c("forest","built","other"),
                   t=c(0,6,14))

summary(obs[[1]])
summary(obs[[1:2]])
}

