% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{FigureOfMerit.plot}
\alias{FigureOfMerit.plot}
\alias{FigureOfMerit.plot,FigureOfMerit-method}
\title{Plot method for FigureOfMerit objects}
\usage{
FigureOfMerit.plot(x, ...)

\S4method{FigureOfMerit.plot}{FigureOfMerit}(x, from, to, type = "b", key,
  scales, xlab, ylab, ...)
}
\arguments{
\item{x}{a \code{FigureOfMerit} object}

\item{...}{additional arguments to \code{xyplot}}

\item{from}{numeric (optional). A single value corresponding to a land use
category. If provided without 'to' the figure of merit for all transitions
from this category will be plotted}

\item{to}{numeric (optional). Similar to 'from'. If provided with a valid
'from' argument the figure of merit for the transition defined by these two
arguments (i.e. from -> to) will be plotted. It is possible to include more
than one category in which case the different transitions will be included
on the same plot}

\item{type}{character. See \code{lattice::\link[lattice]{xyplot}}}

\item{key}{list. See \code{lattice::\link[lattice]{xyplot}}}

\item{scales}{list. See \code{lattice::\link[lattice]{xyplot}}}

\item{xlab}{character or expression. See \code{lattice::\link[lattice]{xyplot}}}

\item{ylab}{character or expression. See \code{lattice::\link[lattice]{xyplot}}}
}
\value{
a trellis object
}
\description{
Plot the overall, category-specific or transition-specific figure of merit at
different resolutions
}
\author{
Simon Moulds
}
\seealso{
\code{\link{FigureOfMerit}}
}

