% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glmModels}
\alias{glmModels}
\title{Model fitting}
\usage{
glmModels(formula, family = binomial, data, obs, model = FALSE, ...)
}
\arguments{
\item{formula}{list containing formula objects}

\item{family}{TODO}

\item{data}{TODO}

\item{obs}{TODO}

\item{model}{TODO}

\item{...}{additional arguments to glm}
}
\value{
TODO
}
\description{
Fit glm models to spatial data
}
\examples{
## TODO
}

