% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{lulccR-package}
\alias{lulccR-package}
\title{lulccR: land use change modelling in R}
\description{
lulccR provides a framework for spatially explicit land use change modelling
in R. The main goal of the package is to allow users to perform every stage of
the modelling process in the same environment. It therefore includes functions
to process and explore model input, fit and evaluate predictive models,
estimate future demand for different land uses, allocate land use change
spatially, calibrate and validate the model, and, finally, visualise model
output.
}
\author{
Simon Moulds
}

