% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Model-class}
\alias{Model-class}
\title{Virtual class Model}
\description{
A virtual S4 class to represent land use change models.
}
\section{Slots}{

\describe{
\item{\code{obs}}{an ObsLulcMaps object}

\item{\code{ef}}{an ExpVarMaps object}

\item{\code{models}}{a PredModels object}

\item{\code{time}}{numeric vector of timesteps over which simulation will occur}

\item{\code{demand}}{matrix containing demand scenario or NULL}

\item{\code{hist}}{RasterLayer showing land use history or NULL}

\item{\code{mask}}{RasterLayer showing masked areas or NULL}

\item{\code{neighb}}{NeighbMaps object or NULL}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}

\item{\code{output}}{RasterStack containing simulated land use maps or NULL}
}}

