% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ModelInput-class}
\alias{CluesModelInput-class}
\alias{ModelInput-class}
\alias{OrderedModelInput-class}
\title{Class ModelInput}
\description{
An S4 class to represent common inputs to land use change models.
}
\section{Slots}{

\describe{
\item{\code{obs}}{RasterLayer showing initial land use}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}

\item{\code{pred}}{a PredictorMaps object}

\item{\code{models}}{a StatModels object}

\item{\code{time}}{numeric vector containing timesteps over which simulation will
occur}

\item{\code{demand}}{matrix containing demand scenario or NULL}

\item{\code{hist}}{RasterLayer showing land use history or NULL}

\item{\code{mask}}{RasterLayer showing masked areas or NULL}

\item{\code{neighb}}{NeighbMaps object or NULL}

\item{\code{rules}}{matrix with land use change decision rules}

\item{\code{nb.rules}}{numeric with neighbourhood decision rules}

\item{\code{elas}}{numeric indicating elasticity to change (only required for
\code{CluesModelInput} objects}

\item{\code{params}}{list with model parameters}
}}
\author{
Simon Moulds
}

