% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{NeighbMaps-class}
\alias{NeighbMaps-class}
\title{Class NeighbMaps}
\description{
An S4 class for neighbourhood maps.
}
\section{Slots}{

\describe{
\item{\code{filename}}{TODO}

\item{\code{layers}}{TODO}

\item{\code{title}}{TODO}

\item{\code{extent}}{TODO}

\item{\code{rotated}}{TODO}

\item{\code{rotation}}{TODO}

\item{\code{ncols}}{TODO}

\item{\code{nrows}}{TODO}

\item{\code{crs}}{TODO}

\item{\code{history}}{TODO}

\item{\code{z}}{TODO}

\item{\code{calls}}{list}

\item{\code{categories}}{numeric vector of land use categories for which neighbourhood
  maps exist}
}}

