% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ObservedMaps}
\alias{ObservedMaps}
\alias{ObservedMaps,RasterLayer,ANY-method}
\alias{ObservedMaps,RasterStack,ANY-method}
\alias{ObservedMaps,character,character-method}
\alias{ObservedMaps,list,character-method}
\alias{ObservedMaps,missing,character-method}
\title{Create an ObservedMaps object}
\usage{
ObservedMaps(x, pattern, ...)

\S4method{ObservedMaps}{missing,character}(x, pattern, ...)

\S4method{ObservedMaps}{character,character}(x, pattern, ...)

\S4method{ObservedMaps}{list,character}(x, pattern, ...)

\S4method{ObservedMaps}{RasterLayer,ANY}(x, pattern, ...)

\S4method{ObservedMaps}{RasterStack,ANY}(x, pattern, categories, labels, t)
}
\arguments{
\item{x}{path (character), Raster* object or list of Raster* objects. Default
behaviour is to search for files in the working directory}

\item{pattern}{regular expression (character). Only filenames (if \code{x} is
a path) or Raster* objects (if \code{x} is a list) matching the regular
expression will be returned. See \code{raster::\link[raster]{raster}} for
more information about supported filetypes}

\item{categories}{numeric vector of land use categories in observed maps}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}}

\item{t}{numeric vector containing the timestep of each observed map. The
first timestep must be 0}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
An ObservedMaps object.
}
\description{
Methods to create an ObservedMaps object, which may be created from file, an
existing Raster* object or a list of Raster* objects.
}
\details{
Observed land use maps should have the same extent and resolution. The
location of non-NA cells in \code{ObservedMaps} objects defines the region for
subsequent analysis.
}
\examples{
## Plum Islands Ecosystem
obs <- ObservedMaps(x=pie,
                    pattern="lu",
                    categories=c(1,2,3),
                    labels=c("forest","built","other"),
                    t=c(0,6,14))

## Sibuyan Island
obs <- ObservedMaps(x=sibuyan,
                    pattern="lu",
                    categories=c(1,2,3,4,5),
                    labels=c("forest","coconut","grass","rice","other"),
                    t=c(0))
}

