% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ObsLulcMaps-class}
\alias{ObsLulcMaps-class}
\title{Class ObsLulcMaps}
\description{
An S4 class for observed land use maps.
}
\section{Slots}{

\describe{
\item{\code{filename}}{TODO}

\item{\code{layers}}{TODO}

\item{\code{title}}{TODO}

\item{\code{extent}}{TODO}

\item{\code{rotated}}{TODO}

\item{\code{rotation}}{TODO}

\item{\code{ncols}}{TODO}

\item{\code{nrows}}{TODO}

\item{\code{crs}}{TODO}

\item{\code{history}}{TODO}

\item{\code{z}}{TODO}

\item{\code{t}}{numeric vector with timesteps corresponding to each observed map}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

