% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ObsLulcMaps}
\alias{ObsLulcMaps}
\alias{ObsLulcMaps,RasterLayer,ANY-method}
\alias{ObsLulcMaps,RasterStack,ANY-method}
\alias{ObsLulcMaps,character,character-method}
\alias{ObsLulcMaps,list,character-method}
\alias{ObsLulcMaps,missing,character-method}
\title{Create an ObsLulcMaps object}
\usage{
ObsLulcMaps(x, pattern, ...)

\S4method{ObsLulcMaps}{missing,character}(x, pattern, ...)

\S4method{ObsLulcMaps}{character,character}(x, pattern, ...)

\S4method{ObsLulcMaps}{list,character}(x, pattern, ...)

\S4method{ObsLulcMaps}{RasterLayer,ANY}(x, pattern, ...)

\S4method{ObsLulcMaps}{RasterStack,ANY}(x, pattern, categories, labels, t)
}
\arguments{
\item{x}{path (character), Raster* object or list of Raster* objects. Default
behaviour is to search for files in the working directory}

\item{pattern}{regular expression (character). Only filenames (if \code{x} is
a path) or Raster* objects (if \code{x} is a list) matching the regular
expression will be returned. See \cr
\code{raster::\link[raster]{raster}} for more information about supported filetypes}

\item{categories}{numeric vector of land use categories in observed maps}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}}

\item{t}{numeric vector containing the timestep of each observed map. The
first timestep must be 0}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
An ObsLulcMaps object.
}
\description{
Methods to create an ObsLulcMaps object, which may be created from file, an
existing Raster* object or a list of Raster* objects.
}
\details{
Observed land use maps should have the same extent and resolution. The
location of non-NA cells in \code{ObsLulcMaps} objects defines the region for
subsequent analysis.
}
\examples{
## Plum Island Ecosystems
obs <- ObsLulcMaps(x=pie,
                   pattern="lu",
                   categories=c(1,2,3),
                   labels=c("forest","built","other"),
                   t=c(0,6,14))

## Sibuyan Island
obs <- ObsLulcMaps(x=sibuyan$maps,
                   pattern="lu",
                   categories=c(1,2,3,4,5),
                   labels=c("forest","coconut","grass","rice","other"),
                   t=c(0,14))
}
\seealso{
\code{\link{ObsLulcMaps-class}}, \code{raster::\link[raster]{stack}}
}

