% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{OrderedModel}
\alias{OrderedModel}
\alias{OrderedModel,ModelInput-method}
\title{Create an OrderedModel object}
\usage{
OrderedModel(x, ...)

\S4method{OrderedModel}{ModelInput}(x, models, hist, mask, neighb = NULL,
  rules = NULL, nb.rules = NULL, order, params, output = NULL, ...)
}
\arguments{
\item{x}{a ModelInput object}

\item{models}{a PredModels object}

\item{hist}{RasterLayer containing land use history (values represent the
number of years the cell has contained the current land use category)}

\item{mask}{RasterLayer containing binary values where 0 indicates cells
that are not allowed to change}

\item{neighb}{an object of class NeighbMaps}

\item{rules}{matrix with land use change decision rules}

\item{nb.rules}{numeric with neighbourhood decision rules}

\item{order}{numeric vector of land use categories in the order that change
should be allocated. See Details}

\item{params}{list with model parameters}

\item{output}{either a RasterStack containing output maps or NULL}

\item{\dots}{additional arguments (none)}
}
\value{
An OrderedModel object.
}
\description{
Methods to create a \code{OrderedModel} object to supply to
\code{\link{allocate}}.
}
\details{
The \code{params} argument is a list of parameter values which should contain
the following components:

\describe{
  \item{\code{max.diff}}{The maximum allowed difference between allocated and
    demanded area of any land use type. Default is 5}
}
}
\examples{
## see lulccR-package examples
}
\references{
Fuchs, R., Herold, M., Verburg, P.H., and Clevers, J.G.P.W. (2013). A
high-resolution and harmonized model approach for reconstructing and analysing
historic land changes in Europe, Biogeosciences, 10:1543-1559.
}
\seealso{
\code{\link{OrderedModel-class}}, \code{\link{allocate}}
}

