% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{OrderedModel2-class}
\alias{OrderedModel2-class}
\title{Class OrderedModel2}
\description{
An S4 class to represent inputs to [...]
}
\details{
TODO
}
\section{Slots}{

\describe{
\item{\code{obs}}{an ObservedMaps object}

\item{\code{pred}}{a PredictorMaps object}

\item{\code{models}}{a StatModels object}

\item{\code{time}}{numeric vector containing timesteps over which simulation will
occur}

\item{\code{demand}}{matrix containing demand scenario or NULL}

\item{\code{hist}}{RasterLayer showing land use history or NULL}

\item{\code{mask}}{RasterLayer showing masked areas or NULL}

\item{\code{neighb}}{NeighbMaps object or NULL}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}

\item{\code{rules}}{matrix with land use change decision rules}

\item{\code{nb.rules}}{numeric with neighbourhood decision rules}

\item{\code{order}}{numeric vector of land use categories in the order that change
should be allocated}

\item{\code{params}}{list with model parameters}

\item{\code{output}}{RasterStack containing simulated land use maps or NULL}
}}

