% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partition}
\alias{partition}
\title{Partition raster data}
\usage{
partition(x, size = 0.5, spatial = TRUE, ...)
}
\arguments{
\item{x}{RasterLayer with categorical data}

\item{size}{numeric value between zero and one indicating the proportion of
non-NA cells that should be included in the training partition. Default is
0.5, which results in equally sized partitions}

\item{spatial}{logical. If TRUE, the function returns a SpatialPoints object
with the coordinates of cells in each partition. If FALSE, the cell numbers
are returned}

\item{\dots}{additional arguments (none)}
}
\value{
A list containing the following components:
\describe{
  \item{\code{train}}{a SpatialPoints object or numeric vector indicating the
  cells in the training partition}
  \item{\code{test}}{a SpatialPoints object or numeric vector indicating the
  cells in the testing partition}
  \item{\code{all}}{a SpatialPoints object or numeric vector indicating all
  non-NA cells in the study region}
}
}
\description{
Divide a categorical raster map into training and testing partitions.
A wrapper function for \cr
\code{caret::\link[caret]{createDataPartition}} (Kuhn, 2008) to divide a
categorical raster map into training and testing partitions.
}
\examples{
\dontrun{

## Plum Island Ecosystems

## observed maps
obs <- ObsLulcMaps(x=pie,
                   pattern="lu",
                   categories=c(1,2,3),
                   labels=c("Forest","Built","Other"),
                   t=c(0,6,14))

## explanatory variables
ef <- ExpVarMaps(x=pie, pattern="ef")

## prepare model input
input <- ModelInput(obs=obs,
                    ef=ef,
                    time=0:14)

part <- partition(x=obs[[1]], size=0.1, spatial=TRUE)
train.data <- as.data.frame(x=input, cells=part[["train"]], t=0)
test.data  <- as.data.frame(x=input, cells=part[["test"]], t=0)

}
}
\references{
Kuhn, M. (2008). Building predictive models in R using the caret
package. Journal of Statistical Software, 28(5), 1-26.
}
\seealso{
\code{caret::\link[caret]{createDataPartition}}
}

