% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Performance-class}
\alias{Performance-class}
\title{Class Performance}
\description{
An S4 class that extends \code{ROCR::\link[ROCR]{performance-class}} to hold
the results of multiple model evaluations.
}
\section{Slots}{

\describe{
\item{\code{performance}}{list of ROCR performance objects. Each object is
calculated for the corresponding ROCR prediction object held in the
Prediction object supplied to the constructor function}

\item{\code{auc}}{TODO}

\item{\code{categories}}{numeric vector of land use categories for which performance
objects were created}

\item{\code{labels}}{character vector with labels corresponding to \code{categories}}
}}

