% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Performance}
\alias{Performance}
\title{Create a Performance object}
\usage{
Performance(pred, measure, x.measure = "cutoff", ...)
}
\arguments{
\item{pred}{an object of class Prediction}

\item{measure}{performance measure to use for the evaluation. See
\code{ROCR::\link[ROCR]{performance}}}

\item{x.measure}{a second performance measure. See
\code{ROCR::\link[ROCR]{performance}}}

\item{\dots}{additional arguments to \code{ROCR::\link[ROCR]{performance}}}
}
\value{
A \code{Performance} object.
}
\description{
This function uses different measures to evaluate multiple
\code{ROCR::\link[ROCR]{prediction}} objects stored in a
\code{\link{Prediction}} object.
}
\examples{
## see PredModels examples
}
\references{
Sing, T., Sander, O., Beerenwinkel, N., Lengauer, T. (2005).
ROCR: visualizing classifier performance in R. Bioinformatics
21(20):3940-3941.
}
\seealso{
\code{\link{performance}}, \code{\link{Prediction}}
}

