% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{PerformanceMulti-class}
\alias{PerformanceMulti-class}
\title{Class PerformanceMulti}
\description{
An S4 class that extends \code{ROCR::\link[ROCR]{performance-class}} to hold
the results of multiple model evaluations.
}
\section{Slots}{

\describe{
\item{\code{performance}}{list of \code{performance} objects. Each object is
calculated for the corresponding \code{prediction} object held in the
\code{\linkS4class{PredictionMulti}} object supplied to the constructor
function}

\item{\code{auc}}{numeric vector containing the area under the ROC curve for each
performance object}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

