% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PerformanceMulti}
\alias{PerformanceMulti}
\title{Create a PerformanceMulti object}
\usage{
PerformanceMulti(pred, measure, x.measure = "cutoff", ...)
}
\arguments{
\item{pred}{an object of class \code{PredictionMulti}}

\item{...}{additional arguments to \code{\link{performance}}}
}
\value{
a \code{PerformanceMulti} object
}
\description{
This function uses different measures to evaluate multiple
\code{ROCR::\link[ROCR]{prediction}} objects stored in a
\code{\link{PredictionMulti}} object.
}
\author{
Simon Moulds
}
\references{
Sing, T., Sander, O., Beerenwinkel, N., Lengauer, T. (2005).
ROCR: visualizing classifier performance in R. Bioinformatics
21(20):3940-3941.
}
\seealso{
\code{\link{performance}}, \code{\link{PredictionMulti}}
}

