% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.agreement,AgreementBudget-method}
\alias{plot.agreement,AgreementBudget-method}
\title{Plot method for AgreementBudget objects}
\usage{
\S4method{plot.agreement}{AgreementBudget}(x, pal = "Set2", key, scales, xlab,
  ylab, ...)
}
\arguments{
\item{x}{object of class \code{AgreementBudget}}

\item{pal}{character. A colour palette. Default is 'Set2' from RColorBrewer}

\item{...}{additional arguments to \code{xyplot}}
}
\value{
a trellis object
}
\description{
Plot sources of agreement and disagreement between three maps at multiple
resolutions
}
\details{
The plot layout is based on work presented in Pontius et al. (2011)
}
\author{
Simon Moulds
}
\references{
Pontius Jr, R.G., Peethambaram, S., Castella, J.C. (2011).
Comparison of three maps at multiple resolutions: a case study of land change
simulation in Cho Don District, Vietnam. Annals of the Association of American
Geographers 101(1): 45-62.
}
\seealso{
\code{\link{AgreementBudget}}
}

