% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.fom,FigureOfMerit-method}
\alias{plot.fom,FigureOfMerit-method}
\title{Plot method for FigureOfMerit objects}
\usage{
\S4method{plot.fom}{FigureOfMerit}(x, from, to, type = "b", key, scales, xlab,
  ylab, ...)
}
\arguments{
\item{x}{a \code{FigureOfMerit} object}

\item{from}{numeric (optional). A single value corresponding to a land use
category. If provided without 'to' the figure of merit for all transitions
from this category will be plotted}

\item{to}{numeric (optional). Similar to 'from'. If provided with a valid
'from' argument the figure of merit for the transition defined by these two
arguments (i.e. from -> to) will be plotted. It is possible to include more
than one category in which case the different transitions will be included
on the same plot}

\item{...}{additional arguments to \code{xyplot}}
}
\value{
a trellis object
}
\description{
Plot the overall, category-specific or transition-specific figure of merit at
different resolutions
}
\author{
Simon Moulds
}
\seealso{
\code{\link{FigureOfMerit}}
}

