% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.roc,PerformanceMulti-method}
\alias{plot.roc,PerformanceMulti-method}
\alias{plot.roc,list-method}
\title{Plot methods for PerformanceMulti objects}
\usage{
\S4method{plot.roc}{PerformanceMulti}(perf, type, ...)

\S4method{plot.roc}{list}(perf, multipanel = TRUE, type = "l",
  abline = list(c(0, 1), col = "grey"), col = RColotBrewer::brewer.pal(9,
  "Set1"), key.args = NULL, ...)
}
\arguments{
\item{perf}{either a single \code{PerformanceMulti} object or a list of these.
If a list is provided it must be named.}

\item{...}{additional arguments to \code{xyplot}}

\item{multipanel}{logical. If \code{TRUE}, create a trellis plot where the
number of panels equals the number of \code{PerformanceMulti} objects.
Otherwise, create a single plot for each \code{PerformanceMulti} object}
}
\value{
a single plot (if \code{multipanel=FALSE}) or a list of plots.
}
\description{
Plot the the ROC curve for each \code{performance} object in a
\code{\link{PerformanceMulti}} object. If a list of \code{PerformanceMulti}
objects is provided, with each list element representing a different type
of mathematical model, ROC curves for different models are included on the
same plot for model comparison
}
\author{
Simon Moulds
}
\seealso{
\code{\link{PerformanceMulti}}
}

