% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Prediction-class}
\alias{Prediction-class}
\title{Class Prediction}
\description{
An S4 class that extends \code{ROCR::\link[ROCR]{prediction-class}} to hold
the results of multiple model predictions.
}
\section{Slots}{

\describe{
\item{\code{prediction}}{a list of \code{ROCR::\link[ROCR]{prediction-class}} objects.
These objects are calculated for each statistical model in the
\code{\link{PredModels}} object supplied to the constructor function}

\item{\code{types}}{character vector of model types for which \code{prediction}
objects were created}

\item{\code{categories}}{numeric vector of land use categories for which
\code{prediction} objects were created}

\item{\code{labels}}{character vector with labels corresponding to \code{categories}}
}}

