% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Prediction}
\alias{Prediction}
\title{Create a Prediction object}
\usage{
Prediction(models, obs, ef, timestep = 0, partition, ...)
}
\arguments{
\item{models}{a \code{PredModels} object}

\item{obs}{an \code{ObsLulcMaps} object}

\item{ef}{an \code{ExpVarMaps} object}

\item{timestep}{numeric indicating the timestep of the observed map in
\code{obs} against which the observed map should be tested}

\item{partition}{index of cells for which occurrence should be predicted}

\item{\dots}{additional arguments to \code{ROCR::\link[ROCR]{prediction}}}
}
\value{
A \code{Prediction} object.
}
\description{
This function creates a \code{ROCR::\link[ROCR]{prediction}} object for each
predictive model in a \code{PredModels} object. It should be used with
\code{\link{Performance}} to evaluate multiple models with exactly the
same criteria while keeping track of which model corresponds to which land use
category. This makes it easier to write functions that compare different model
types for the same land use category, such as \code{\link{Performance.plot}}.
}
\references{
Sing, T., Sander, O., Beerenwinkel, N., Lengauer, T. (2005).
ROCR: visualizing classifier performance in R. Bioinformatics
21(20):3940-3941.
}
\seealso{
\code{ROCR::\link[ROCR]{prediction}}
}

