% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{PredictorCall,call,character,character-method}
\alias{PredictorCall,PredictorCall,RasterLayer,ANY-method}
\alias{PredictorCall,call,character,character-method}
\title{Create a PredictorCall object}
\usage{
\S4method{PredictorCall}{call,character,character}(call, update.arg, name, ...)

\S4method{PredictorCall}{PredictorCall,RasterLayer,ANY}(call, update.arg, name,
  ...)
}
\arguments{
\item{call}{an unevaluated function call}

\item{update.arg}{argument (character) to \code{call} that should be
updated when a new land use map is available during simulation}

\item{name}{name (character) of the predictor variable}

\item{...}{additional arguments (none)}
}
\value{
PredictorCall object
}
\description{
Methods to create a PredictorCall object.
}
\examples{
## Plum Island Ecosystem
obs <- ObservedMaps(x=pie,
                    pattern="lu",
                    labels=c("forest","built","other"),
                    t=c(0,6,14))

## function to calculate distance to land use category
myfun <- function(x, category, ...) {
   y <- x
   y[!is.na(y)] <- 1
   x[x != value] <- NA
   dist <- distance(x, ...)
   dist <- dist * y
}

## PredictorCall object for distance to forest
dist2forest <- PredictorCall(call=call("myfun", x=obs@maps[[1]], category=1),
                             update.arg="x",
                             name="dist2forest")
}
\author{
Simon Moulds
}
\seealso{
\code{\link{Predictors}}
}

