% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predictorMaps}
\alias{predictorMaps}
\alias{predictorMaps,RasterStack,character-method}
\alias{predictorMaps,character,character-method}
\alias{predictorMaps,list,character-method}
\alias{predictorMaps,missing,character-method}
\title{Load predictor variables}
\usage{
predictorMaps(x, pattern, ...)

\S4method{predictorMaps}{missing,character}(x, pattern, ...)

\S4method{predictorMaps}{character,character}(x, pattern, ...)

\S4method{predictorMaps}{list,character}(x, pattern, ...)

\S4method{predictorMaps}{RasterStack,character}(x, pattern, ...)
}
\arguments{
\item{x}{path (character) to directory containing observed land use maps,
a Raster* object or a list of Raster* objects.}

\item{pattern}{regular expression (character). Only filenames (if \code{x} is
a path) or Raster* objects (if \code{x} is a list) matching the regular
expression will be returned. See \code{raster::\link{raster}} for more
information about supported filetypes}

\item{...}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
a predictorMaps object
}
\description{
Methods to load maps of predictor variables, which may be created from file, an
existing Raster* object or a list of Raster* objects.
}
\details{
Predictor maps should follow a naming convention to identify them as static
(one map provided for the study period) or dynamic (one map provided for each
year of the study period). The name should consist of two (static) or three
(dynamic) parts: firstly, the prefix should differentiate predictor maps from
other maps in the directory, list or RasterStack. This should be followed
by a unique number to differentiate the predictor maps (note that the order of
predictor variables in the predictorMaps object is determined by this value)
If the predictor is dynamic this number should be followed by a second number
representing the timestep to which the map applies. Dynamic variables should
include a map for time 0 (corresponding to the initial observed map) and every
subsequent timestep. The different parts should be separated by a period or
underscore.

Maps of different predictor variables should have the same coordinate
reference system but do not have to have the same extent and resolution as
long as the minimum extent is that of the study region defined by an
\code{ObservedMaps} object. However, maps for different timesteps of the same
dynamic predictor variable should have the same extent and resolution because
these are stored as RasterStack objects.
}
\examples{
## Plum Island Ecosystem
pred.maps <- PredictorMaps(x=pie, pattern="pred")

## Sibuyan
pred.maps <- PredictorMaps(x=sibuyan, pattern="pred")
}
\author{
Simon Moulds
}
\seealso{
\code{\link{Predictors}}
}

