% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Predictors}
\alias{Predictors}
\alias{Predictors,list,list-method}
\alias{Predictors,list,missing-method}
\alias{Predictors,missing,list-method}
\title{Create a Predictors object}
\usage{
Predictors(maps, calls, ...)

\S4method{Predictors}{missing,list}(maps, calls, ...)

\S4method{Predictors}{list,missing}(maps, calls, ...)

\S4method{Predictors}{list,list}(maps, calls, ...)
}
\arguments{
\item{maps}{a list of RasterStack objects such as that returned by
\code{\link{predictorMaps}}}

\item{calls}{a list of PredictorCall objects}

\item{...}{additional arguments (none)}
}
\value{
Predictors object
}
\description{
Methods to create a Predictors object.
}
\examples{
## function to calculate distance to land use category
myfun <- function(x, category, ...) {
   y <- x
   y[!is.na(y)] <- 1
   x[x != value] <- NA
   dist <- distance(x, ...)
   dist <- dist * y
}

## PredictorCall object for distance to forest
call <- PredictorCall(call=call("myfun", x=obs@maps[[1]], category=1),
                      update.arg="x",
                      name="dist2forest")

maps <- predictorMaps(x=pie, pattern="pred")
pred <- Predictors(maps=maps, calls=list(call))
}
\author{
Simon Moulds
}
\seealso{
\code{\link{PredictorCall}}, \code{\link{predictorMaps}}
}

