% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{PredModels-class}
\alias{PredModels-class}
\title{Class PredModels}
\description{
An S4 class to hold multiple mathematical models for different land use
categories belonging to the same map.
}
\section{Slots}{

\describe{
\item{\code{models}}{list of predictive models}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector with labels corresponding to \code{categories}}
}}

