% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predModelType}
\alias{predModelType}
\alias{predModelType,PredModels-method}
\alias{predModelType,glm-method}
\alias{predModelType,list-method}
\alias{predModelType,randomForest-method}
\alias{predModelType,rpart-method}
\title{Get predictive model type}
\usage{
predModelType(x, ...)

\S4method{predModelType}{glm}(x, ...)

\S4method{predModelType}{rpart}(x, ...)

\S4method{predModelType}{randomForest}(x, ...)

\S4method{predModelType}{PredModels}(x, ...)

\S4method{predModelType}{list}(x, ...)
}
\arguments{
\item{x}{a PredModels object, a model or a list of models}

\item{\dots}{additional arguments (none)}
}
\value{
Character.
}
\description{
This function returns the type of predictive model in use.
}

