% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resample,Predictors,Raster-method}
\alias{resample,Predictors,Raster-method}
\alias{resample,list,Raster-method}
\title{Resample maps in Predictors object or list}
\usage{
\S4method{resample}{Predictors,Raster}(x, y, method = "ngb", ...)

\S4method{resample}{list,Raster}(x, y, method = "ngb", ...)
}
\arguments{
\item{x}{a Predictors object or list of Raster* maps to be resampled}

\item{y}{Raster* object with parameters that \code{x} should be resampled to}

\item{method}{method used to compute values for the new RasterLayer, should be
\code{"bilinear"} for bilinear interpolation, or \code{"ngb"} for nearest
neighbour}

\item{...}{additional arguments to \code{raster::\link[raster]{resample}}}
}
\value{
Predictors object or list, depending on \code{x}
}
\description{
A wrapper function for \code{raster::\link[raster]{resample}} to resample
raster objects in a \link{Predictors-class} object or list.
}
\examples{
## create Predictors object
pred <- predictorMaps(x=pie, pattern="pred")
pred <- Predictors(maps=pred)

## resample to ensure maps have same characteristics as observed maps
pred <- resample(x=pred, y=pie$lu_pie_1985, method="ngb")
}
\author{
Simon Moulds
}
\seealso{
\code{\link{Predictors}}
}

