% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roundSum}
\alias{roundSum}
\title{Round elements in matrix or data.frame rows}
\usage{
roundSum(x, ncell)
}
\arguments{
\item{x}{matrix or data.frame}

\item{ncell}{numeric specifying the target sum for each row in \code{x}}
}
\value{
a matrix
}
\description{
Round all numbers in a matrix or data.frame while ensuring that all rows sum
to the same value
}
\details{
The main application of \code{roundSum} is to ensure that each row in the
demand matrix specifies exactly the number of cells to be allocated to each
land use category for the respective timestep. It may also be used to convert
demand from area to number of cells
}
\author{
Simon Moulds
}

