% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,ExpVarMaps-method}
\alias{show,ExpVarMaps-method}
\alias{show,ModelInput-method}
\alias{show,Performance-method}
\alias{show,PredModels-method}
\alias{show,Prediction-method}
\alias{show,ThreeMapComparison-method}
\title{Show}
\usage{
\S4method{show}{ExpVarMaps}(object)

\S4method{show}{PredModels}(object)

\S4method{show}{Prediction}(object)

\S4method{show}{Performance}(object)

\S4method{show}{ModelInput}(object)

\S4method{show}{ThreeMapComparison}(object)
}
\arguments{
\item{object}{an object belonging to one of the classes in \code{lulcc}}
}
\description{
Show objects
}

