% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{StatModels-class}
\alias{StatModels-class}
\title{Class StatModels}
\description{
An S4 class to hold multiple mathematical models for different land use
categories belonging to the same map.
}
\section{Slots}{

\describe{
\item{\code{models}}{list of mathematical models}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

