% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{StatModels}
\alias{StatModels}
\alias{StatModels,list,ANY,ANY,numeric,character-method}
\alias{StatModels,list,ANY,numeric,character-method}
\alias{StatModels,list,ObservedMaps,ANY,ANY-method}
\title{Create a StatModels object}
\usage{
StatModels(models, obs, categories, labels, ...)

\S4method{StatModels}{list,ObservedMaps,ANY,ANY}(models, obs, categories,
  labels, ...)

\S4method{StatModels}{list,ANY,numeric,character}(models, obs, categories,
  labels, ...)
}
\arguments{
\item{models}{a list of mathematical models}

\item{obs}{an ObservedMaps object}

\item{categories}{numeric vector of land use categories in observed maps. Only
required if \code{obs} is not provided}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}. Only required if \code{obs} is not provided}

\item{...}{additional arguments (none)}
}
\value{
an object of class \code{\link{StatModels}}
}
\description{
Methods to create a \link{StatModels-class} object.
}
\author{
Simon Moulds
}

