% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{subset,ExpVarMaps-method}
\alias{subset,ExpVarMaps-method}
\alias{subset,Performance-method}
\alias{subset,PredModels-method}
\alias{subset,Prediction-method}
\title{Subset}
\usage{
\S4method{subset}{ExpVarMaps}(x, subset, ...)

\S4method{subset}{PredModels}(x, subset, ...)

\S4method{subset}{Performance}(x, subset, ...)

\S4method{subset}{Prediction}(x, subset, ...)
}
\arguments{
\item{x}{TODO}

\item{subset}{TODO}

\item{...}{additional arguments}
}
\value{
TODO
}
\description{
TODO
}
\examples{
## TODO
}

