% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,ExpVarMaps-method}
\alias{summary,ObsLulcMaps-method}
\alias{summary,Performance-method}
\alias{summary,PredModels-method}
\alias{summary,Prediction-method}
\title{Summary}
\usage{
summary(object, ...)

\S4method{summary}{ObsLulcMaps}(object, ...)

\S4method{summary}{ExpVarMaps}(object, ...)

\S4method{summary}{PredModels}(object, ...)

\S4method{summary}{Prediction}(object, ...)

\S4method{summary}{Performance}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Summarize lulcc objects
}
\examples{
## TODO
}

