% Generated by roxygen2 (4.0.2): do not edit by hand
\name{total}
\alias{total}
\title{Total number of cells in a categorical Raster* object}
\usage{
total(x, categories)
}
\arguments{
\item{x}{Raster* object}

\item{categories}{numeric vector containing land use categories. Only cells
  belonging to these categories will be counted}
}
\value{
A list containing the following components:
\describe{
  \item{\code{total}:} a matrix containing the total number of cells belonging
  to each category. Rows represent layers in the input Raster* object
  \item{\code{categories}:} the categories included in the calculation
}
}
\description{
Count the number of cells belonging to each category in a Raster* object.
}
\examples{
# RasterLayer
total(x=sib$lu_sib_1997)

# RasterStack
total(x=stack(pie$lu_pie_1985, pie$lu_pie_1991, pie$lu_pie_1999))
}
\author{
Simon Moulds
}

