% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update.data.frame}
\alias{update.data.frame}
\title{Update data.frame}
\usage{
update.data.frame(x, y, map, cells, timestep, ...)
}
\arguments{
\item{x}{data.frame to be updated}

\item{y}{object of class Predictors}

\item{map}{RasterLayer to be supplied to \code{\link{PredictorCall}} objects}

\item{cells}{index of cells to be extracted. Number of elements must equal the
the number of rows in \code{x}}

\item{timestep}{numeric indicating the timestep under consideration}

\item{...}{additional arguments (none)}
}
\value{
a data.frame
}
\description{
Update a data.frame of predictor variables if dynamic variables or variables
from PredictorCall objects are present.
}
\author{
Simon Moulds
}
\seealso{
\code{\link{as.data.frame}}
}

