% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AgreementBudget}
\alias{AgreementBudget}
\title{Create an AgreementBudget object}
\usage{
AgreementBudget(x, from, to, ...)
}
\arguments{
\item{x}{a \code{ThreeMapComparison} object}

\item{from}{optional numeric value representing a land use category. If
provided, results will be restricted to transitions from this category}

\item{to}{similar to \code{from}. If provided with a valid \code{from}
argument the result will be restricted to the transition defined by
these two arguments (i.e. \code{from} -> \code{to})}

\item{\dots}{additional arguments (none)}
}
\value{
An \code{AgreementBudget} object.
}
\description{
This function quantifies sources of agreement and disagreement between a
reference map for time 1, a reference map for time 2 and a simulated map for
time 2 to provide meaningful information about the performance of land use
change simulations.
}
\details{
The types of agreement and disagreement considered are those descibed in
Pontius et al. (2011):

\enumerate{
  \item Persistence simulated correctly (agreement)
  \item Persistence simulated as change (disagreement)
  \item Change simulated incorrectly (disagreement)
  \item Change simulated correctly (agreement)
  \item Change simulated as persistence (disagreement)
}
}
\examples{
## Sibuyan Island

## load clues.model for Sibuyan Island
sib.clues.model <- sibuyan$intermediate$clues.model

sib.clues.tables <- ThreeMapComparison(x=sib.clues.model,
                                       factors=2^(1:9),
                                       timestep=14)

sib.clues.agr <- AgreementBudget(x=sib.clues.tables)
p <- AgreementBudget.plot(x=sib.clues.agr)
print(p)
}
\references{
Pontius Jr, R.G., Peethambaram, S., Castella, J.C. (2011).
Comparison of three maps at multiple resolutions: a case study of land change
simulation in Cho Don District, Vietnam. Annals of the Association of American
Geographers 101(1): 45-62.
}
\seealso{
\code{\link{AgreementBudget-class}},
\code{\link{AgreementBudget.plot}}, \code{\link{ThreeMapComparison}},
\code{\link{FigureOfMerit}}
}

