% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{allocate}
\alias{allocate}
\alias{allocate,CluesModelInput-method}
\alias{allocate,OrderedModelInput-method}
\title{Allocate land use change spatially}
\usage{
allocate(input, ...)

\S4method{allocate}{CluesModelInput}(input, ...)

\S4method{allocate}{OrderedModelInput}(input, ...)
}
\arguments{
\item{input}{an object inheriting from class \code{ModelInput}}

\item{\dots}{additional arguments for specific methods}
}
\description{
Perform spatially explicit allocation of land use change using different
methods. Currently the function provides an algorithm based on the Change in
Land Use and its Effects at Small regional extent (CLUE-S) model(Verburg et
al., 2002) and a novel stochastic procedure that aims to represent the
uncertainty associated with the allocation of change.
}
\references{
Verburg, P.H., Soepboer, W., Veldkamp, A., Limpiada, R., Espaldon,
V., Mastura, S.S. (2002). Modeling the spatial dynamics of regional land use:
the CLUE-S model. Environmental management, 30(3):391-405.
}
\seealso{
\code{\link{ModelInput}}
}

