% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,PredictorMaps-method}
\title{Coerce \code{PredictorMaps} object to data.frame}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{PredictorMaps}(x, row.names = NULL,
  optional = FALSE, cells, timestep = 0, ...)
}
\arguments{
\item{x}{an object of class \code{\link{PredictorMaps}}}

\item{row.names}{NULL or a character vector giving the row.names for the
data.frame. Missing values are not allowed}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional}

\item{...}{additional arguments (none)}

\item{cells}{index of cells to be extracted}

\item{timestep}{numeric indicating the timestep under consideration. Only
relevant if x@maps contains dynamic predictor variables}
}
\value{
data.frame
}
\description{
This function extracts data from all rasters in a \code{\link{PredictorMaps}}
object for a specified timestep (if dynamic variables are present).
}
\author{
Simon Moulds
}
\seealso{
\code{\link{PredictorMaps}}, \code{\link{partition}}
}

