% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{calcProb}
\alias{calcProb}
\alias{calcProb,StatModels-method}
\alias{calcProb,glm-method}
\alias{calcProb,randomForest-method}
\alias{calcProb,rpart-method}
\title{Estimate location suitability}
\usage{
calcProb(object, ...)

\S4method{calcProb}{glm}(object, newdata, ...)

\S4method{calcProb}{rpart}(object, newdata, ...)

\S4method{calcProb}{randomForest}(object, newdata, ...)

\S4method{calcProb}{StatModels}(object, newdata, df = FALSE, ...)
}
\arguments{
\item{object}{a \code{StatModels} object or a model of any class
for which a predict method exists (currently, \code{glm},
\code{rpart::\link[rpart]{rpart}} and
\code{randomForest::\link[randomForest]{randomForest}}}

\item{newdata}{data.frame containing new data}

\item{df}{logical indicating whether the function should return a matrix
(default) or data.frame}

\item{\dots}{additional arguments to the \code{predict}}
}
\value{
A matrix or data.frame.
}
\description{
Estimate location suitability with predictive models.
}
\details{
This function is usually called from \code{allocate} to calculate land use
suitability at each timestep. However, it may also be used to produce
suitability maps (see examples).
}
\author{
Simon Moulds
}
\seealso{
\code{\link{StatModels}},\code{\link{allocate}},\code{\link{predict}}
}

