% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{calcProb}
\alias{calcProb}
\alias{calcProb,StatModels-method}
\alias{calcProb,glm-method}
\alias{calcProb,randomForest-method}
\alias{calcProb,rpart-method}
\title{Estimate location suitability}
\usage{
calcProb(object, ...)

\S4method{calcProb}{glm}(object, newdata, ...)

\S4method{calcProb}{rpart}(object, newdata, ...)

\S4method{calcProb}{randomForest}(object, newdata, ...)

\S4method{calcProb}{StatModels}(object, newdata, df = FALSE, ...)
}
\arguments{
\item{object}{a \code{StatModels} object or a model of any class
for which a predict method exists (currently, \code{glm},
\code{rpart::\link[rpart]{rpart}} and
\code{randomForest::\link[randomForest]{randomForest}}}

\item{...}{additional arguments to \code{predict}}

\item{newdata}{data.frame containing new data}

\item{df}{logical indicating whether the function should return a data.frame
or matrix (default)}
}
\value{
a data.frame or matrix
}
\description{
Estimate location suitability with predictive models.
}
\author{
Simon Moulds
}
\seealso{
\code{\link{StatModels}}, \code{\link{predict}}
}

