% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Extract by index}
\alias{Extract by index}
\alias{[[,CategoryLabel,ANY,ANY-method}
\alias{[[,ExpVarMaps,ANY,ANY-method}
\title{Extract by index}
\usage{
\S4method{[[}{ExpVarMaps,ANY,ANY}(x, i, j, ...)

\S4method{[[}{CategoryLabel,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{TODO}

\item{i}{TODO}

\item{j}{TODO}

\item{...}{additional arguments}
}
\value{
TODO
}
\description{
TODO
}
\examples{
## TODO
}

