% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FigureOfMerit}
\alias{FigureOfMerit}
\title{Create a FigureOfMerit object}
\usage{
FigureOfMerit(x, ...)
}
\arguments{
\item{x}{a \code{ThreeMapComparison} object}

\item{...}{additional arguments (none)}
}
\value{
a \code{FigureOfMerit} object
}
\description{
Calculate the figure of merit at different levels and at different resolutions
for a reference map at time 1, a reference map at time 2 and a simulated map
at time 2.
}
\details{
In land use change modelling terms, the figure of merit is the intersection
of observed change and simulated change divided by the union of these, with
a range of 0 (perfect disagreement) to 1 (perfect agreement). It is useful to
to calculate the figure of merit at three levels: (1) considering all possible
transitions from all land use categories, (2) considering all transitions from
specific land use categories and (3) considering a specific transition from
one land use category to another.
}
\author{
Simon Moulds
}
\references{
Pontius Jr, R.G., Peethambaram, S., Castella, J.C. (2011).
Comparison of three maps at multiple resolutions: a case study of land change
simulation in Cho Don District, Vietnam. Annals of the Association of American
Geographers 101(1): 45-62.
}
\seealso{
\code{\link{ThreeMapComparison}}, \code{\link{AgreementBudget}}
}

