% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glmModels}
\alias{glmModels}
\title{Model fitting}
\usage{
glmModels(formula, family = binomial, data, obs, categories, labels,
  model = FALSE, ...)
}
\arguments{
\item{formula}{list containing formula objects}

\item{family}{see \code{\link[stats]{glm}}. Default is 'binomial'}

\item{data}{see \code{\link[stats]{glm}}}

\item{obs}{an ObsLulcMaps object}

\item{categories}{numeric vector of land use categories in observed maps.
Only required if 'obs' is missing}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}. Only required if 'obs' is missing}

\item{model}{see \code{\link[stats]{glm}}. Default is FALSE}

\item{...}{additional arguments to \code{\link[stats]{glm}}}
}
\value{
A PredModels object.
}
\description{
Fit glm models to spatial data
}
\examples{
## see lulccR-package examples
}
\seealso{
\code{\link[stats]{glm}}, \code{\link{PredModels}}
}

