% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ModelInput-class}
\alias{ModelInput-class}
\title{Class ModelInput}
\description{
An S4 class to represent land use change model inputs.
}
\section{Slots}{

\describe{
\item{\code{obs}}{an ObsLulcMaps object}

\item{\code{ef}}{an ExpVarMaps object}

\item{\code{time}}{numeric vector containing timesteps over which simulation will
occur}

\item{\code{demand}}{matrix containing demand scenario or NULL}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

