% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ModelInput-class}
\alias{ModelInput-class}
\title{Class ModelInput}
\description{
An S4 class to represent common inputs to land use change models.
}
\details{
TODO
}
\section{Slots}{

\describe{
\item{\code{map0}}{RasterLayer showing initial land use}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}

\item{\code{pred}}{a PredictorMaps object}

\item{\code{models}}{a StatModels object}

\item{\code{time}}{numeric vector containing timesteps over which simulation will
occur}

\item{\code{demand}}{matrix containing demand scenario or NULL}

\item{\code{hist}}{RasterLayer showing land use history or NULL}

\item{\code{mask}}{RasterLayer showing masked areas or NULL}

\item{\code{neighb}}{NeighbMaps object or NULL}
}}

