% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ModelInput}
\alias{ModelInput}
\alias{ModelInput,ObsLulcMaps,ExpVarMaps,PredModels,numeric,matrix-method}
\title{Create a ModelInput object}
\usage{
ModelInput(obs, ef, models, time, demand, ...)

\S4method{ModelInput}{ObsLulcMaps,ExpVarMaps,PredModels,numeric,matrix}(obs, ef,
  models, time, demand, hist, mask, neighb = NULL, ...)
}
\arguments{
\item{obs}{an ObsLulcMaps or ModelInput object}

\item{ef}{an ExpVarMaps object}

\item{models}{a PredModels object}

\item{time}{numeric vector containing timesteps over which simulation will
occur}

\item{demand}{matrix with demand for each land use category in terms of number
of cells to be allocated. The first row should be the number of cells
allocated to the initial observed land use map (i.e. the land use map for
time 0)}

\item{hist}{RasterLayer containing land use history (values represent the
number of years the cell has contained the current land use category)}

\item{mask}{RasterLayer containing binary values where 0 indicates cells
that are not allowed to change}

\item{neighb}{an object of class NeighbMaps}

\item{\dots}{additional arguments (none)}
}
\value{
A ModelInput object.
}
\description{
Methods to combine several object classes that are useful for land use change
modelling and perform checks to ensure the objects are compatible in time and
space for a model simulation.
}
\examples{
## see lulccR-package examples
}
\seealso{
\code{\link{ModelInput-class}}
}

