% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ModelInput,ModelInput,ANY,ANY,ANY,ANY-method}
\alias{CluesModelInput,ModelInput,numeric-method}
\alias{CluesModelInput,ObservedMaps,numeric-method}
\alias{ModelInput,ModelInput,ANY,ANY,ANY,ANY,ANY,numeric-method}
\alias{ModelInput,ModelInput,ANY,ANY,ANY,ANY,ANY-method}
\alias{ModelInput,ModelInput,ANY,ANY,ANY,ANY-method}
\alias{ModelInput,ObservedMaps,PredictorMaps,StatModels,numeric,missing,matrix-method}
\alias{ModelInput,ObservedMaps,Predictors,StatModels,numeric,SpatialPoints,matrix-method}
\alias{ModelInput,ObservedMaps,Predictors,StatModels,numeric,matrix-method}
\title{Create a ModelInput object}
\usage{
\S4method{ModelInput}{ModelInput,ANY,ANY,ANY,ANY}(x, pred, models, time, demand,
  ...)

\S4method{ModelInput}{ObservedMaps,Predictors,StatModels,numeric,matrix}(x,
  pred, models, time, demand, hist, mask, neighb = NULL, ...)

\S4method{CluesModelInput}{ModelInput,numeric}(x, elas, rules = NULL,
  nb.rules = NULL, params, ...)

\S4method{CluesModelInput}{ObservedMaps,numeric}(x, elas, rules = NULL,
  nb.rules = NULL, params, ...)
}
\arguments{
\item{pred}{a Predictors object}

\item{models}{a StatModels object}

\item{time}{numeric vector containing timesteps over which simulation will
occur}

\item{demand}{matrix with demand for each land use category in terms of number
of cells to be allocated. The first row should be the number of cells
allocated to the initial observed land use map (i.e. the land use map for
time 0)}

\item{...}{additional arguments (none)}

\item{hist}{RasterLayer containing land use history (values represent the
number of years the cell has contained the current land use category)}

\item{mask}{RasterLayer containing binary values where 0 indicates cells
that are not allowed to change}

\item{neighb}{an object of class NeighbMaps}

\item{elas}{numeric indicating elasticity of each land use category to change
(only required for \code{CluesModelInput} objects}

\item{rules}{matrix with land use change decision rules}

\item{nb.rules}{numeric with neighbourhood decision rules}

\item{params}{list with model parameters}

\item{obs}{an ObservedMaps object}

\item{points}{SpatialPoints* object defining the points over which simulation
will occur}

\item{checkExtent}{logical. Should the input maps be checked for NA values
in the study region defined by \code{points}?}
}
\value{
a ModelInput object
}
\description{
Methods to combine several object classes that are useful for land use change
modelling and perform a series of checks to ensure the objects are compatible
in time and space for a model simulation. If they are, a ModelInput object is
created which may be supplied to \code{\link{allocate}}.
}
\author{
Simon Moulds
}
\seealso{
\code{\link{allocate}}
}

