% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ModelInput}
\alias{ModelInput}
\alias{ModelInput,ObsLulcMaps,ExpVarMaps-method}
\title{Create a ModelInput object}
\usage{
ModelInput(obs, ef, ...)

\S4method{ModelInput}{ObsLulcMaps,ExpVarMaps}(obs, ef, time, ...)
}
\arguments{
\item{obs}{an ObsLulcMaps or ModelInput object}

\item{ef}{an ExpVarMaps object}

\item{time}{numeric vector containing timesteps over which simulation will
occur}

\item{\dots}{additional arguments (none)}
}
\value{
A ModelInput object.
}
\description{
Methods to combine several object classes that are useful for land use change
modelling and perform checks to ensure the objects are compatible in time and
space for a model simulation.
}
\examples{
## see lulccR-package examples
}
\seealso{
\code{\link{ModelInput-class}}
}

