% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ModelInput}
\alias{CluesModelInput,ModelInput,numeric-method}
\alias{CluesModelInput,ObservedMaps,numeric-method}
\alias{ModelInput}
\alias{ModelInput,ModelInput,ANY,ANY,ANY,ANY-method}
\alias{ModelInput,ObservedMaps,Predictors,StatModels,numeric,matrix-method}
\alias{OrderedModelInput,ModelInput-method}
\alias{OrderedModelInput,ObservedMaps-method}
\title{Create a ModelInput object}
\usage{
ModelInput(x, pred, models, time, demand, ...)

\S4method{ModelInput}{ModelInput,ANY,ANY,ANY,ANY}(x, pred, models, time, demand,
  ...)

\S4method{ModelInput}{ObservedMaps,Predictors,StatModels,numeric,matrix}(x,
  pred, models, time, demand, hist, mask, neighb = NULL, ...)

\S4method{CluesModelInput}{ModelInput,numeric}(x, elas, rules = NULL,
  nb.rules = NULL, params, ...)

\S4method{CluesModelInput}{ObservedMaps,numeric}(x, elas, rules = NULL,
  nb.rules = NULL, params, ...)

\S4method{OrderedModelInput}{ModelInput}(x, rules = NULL, nb.rules = NULL,
  params, ...)

\S4method{OrderedModelInput}{ObservedMaps}(x, rules = NULL, nb.rules = NULL,
  params, ...)
}
\arguments{
\item{x}{an ObservedMaps object or a ModelInput object}

\item{pred}{a Predictors object}

\item{models}{a StatModels object}

\item{time}{numeric vector containing timesteps over which simulation will
occur}

\item{demand}{matrix with demand for each land use category in terms of number
of cells to be allocated. The first row should be the number of cells
allocated to the initial observed land use map (i.e. the land use map for
time 0)}

\item{...}{additional arguments (none)}

\item{hist}{RasterLayer containing land use history (values represent the
number of years the cell has contained the current land use category)}

\item{mask}{RasterLayer containing binary values where 0 indicates cells
that are not allowed to change}

\item{neighb}{an object of class NeighbMaps}

\item{elas}{numeric indicating elasticity of each land use category to change
(only required for \code{CluesModelInput} objects}

\item{rules}{matrix with land use change decision rules}

\item{nb.rules}{numeric with neighbourhood decision rules}

\item{params}{list with model parameters}
}
\value{
a ModelInput object
}
\description{
Methods to combine several object classes that are useful for land use change
modelling and perform a series of checks to ensure the objects are compatible
in time and space for a model simulation. If they are, a ModelInput object is
created which may be supplied to \code{\link{allocate}}.
}
\details{
The \code{params} argument is a list of parameter values. For
\code{CluesModelInput} it should contain the following components:
\describe{
  \item{\code{jitter.f}}{Parameter controlling the amount of perturbation
    applied to the probability surface prior to running the CLUE-S iterative
    algorithm. Higher values result in more perturbation. Default is 0.0001}
  \item{\code{scale.f}}{In CLUE-S the suitability of land use types that do
    not meet demand is increased by an amount obtained by multiplying this
    parameter by the difference between allocated and demanded area. Default
    is 0.0005}
  \item{\code{max.iter}}{The maximum number of iterations in the simulation}
  \item{\code{max.diff}}{The maximum allowed difference between allocated and
    demanded area of any land use type. Default is 5}
  \item{\code{ave.diff}}{The average allowed difference between allocated and
    demanded area. Default is 5}
}

When \code{params} is passed to \code{OrderedModelInput} it should only
contain \code{max.diff}, which has the same meaning as for
\code{CluesModelInput}.
}
\author{
Simon Moulds
}
\seealso{
\code{\link{allocate}}
}

