% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{NeighbMaps-class}
\alias{NeighbMaps-class}
\title{Class NeighbMaps}
\description{
An S4 class for neighbourhood maps.
}
\section{Slots}{

\describe{
\item{\code{filename}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{layers}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{title}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{extent}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotated}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotation}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{ncols}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{nrows}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{crs}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{history}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{z}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{calls}}{list containing each call to \code{raster::\link[raster]{focal}}}

\item{\code{categories}}{numeric vector of land use categories for which neighbourhood
  maps exist}
}}

