% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{NeighbMaps-class}
\alias{NeighbMaps-class}
\title{Class NeighbMaps}
\description{
An S4 class for neighbourhood maps.
}
\section{Slots}{

\describe{
\item{\code{maps}}{list of RasterLayers showing neighbourhood values for each land use in
\code{categories}}

\item{\code{categories}}{numeric vector of land use categories for which neighbourhood
maps were calculated}

\item{\code{weights}}{list of weights matrices}

\item{\code{fun}}{function used to calculate neighbourhood values}

\item{\code{focal.args}}{list of all other arguments supplied to
  \code{raster::\link[raster]{focal}}}
}}

