% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{OrderedModel}
\alias{OrderedModel}
\alias{OrderedModel,ModelInput-method}
\title{Create an OrderedModel object}
\usage{
OrderedModel(x, ...)

\S4method{OrderedModel}{ModelInput}(x, rules = NULL, nb.rules = NULL, order,
  params, output = NULL, ...)
}
\arguments{
\item{x}{an ObservedMaps object or a ModelInput object}

\item{rules}{matrix with land use change decision rules}

\item{nb.rules}{numeric with neighbourhood decision rules}

\item{order}{numeric vector of land use categories in the order that change
should be allocated. See Details}

\item{params}{list with model parameters}

\item{output}{either a RasterStack containing output maps or NULL}

\item{\dots}{additional arguments to \code{\link{ModelInput}}}
}
\value{
An OrderedModel object.
}
\description{
Methods to create a \code{OrderedModel} object to supply to
\code{\link{allocate}}.
}
\details{
The \code{params} argument is a list of parameter values which should contain
the following components:

\describe{
  \item{\code{max.diff}}{The maximum allowed difference between allocated and
    demanded area of any land use type. Default is 5}
}

An ordered allocation procedure is frequently used in large study areas TODO
}
\seealso{
\code{link{ModelInput}},\code{\link{allocate}}
}

