% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partition}
\alias{partition}
\title{Partition raster data}
\usage{
partition(x, size = 0.5, spatial = TRUE, ...)
}
\arguments{
\item{x}{RasterLayer with categorical data}

\item{size}{numeric between zero and one indicating the proportion of non-NA
cells that should be included in the training partition. Default is 0.5,
which results in equally sized partitions}

\item{spatial}{logical. If TRUE, the function returns a SpatialPoints object
with the coordinates of cells in each partition. If FALSE, the cell numbers
are returned}

\item{...}{additional arguments (none)}
}
\value{
a list containing the following components:
\describe{
  \item{\code{train}}{a SpatialPoints object or numeric vector indicating the
  cells in the training partition}
  \item{\code{test}}{a SpatialPoints object or numeric vector indicating the
  cells in the testing partition}
}
}
\description{
Divide a categorical raster map into training and testing partitions.
}
\examples{
obs <- ObservedMaps(x=sibuyan,
                    pattern="lu",
                    categories=c(1,2,3,4,5),
                    labels=c("forest","coconut","grass","rice","other"),
                    t=c(0))

## create equally sized training and testing partitions
partition(x=obs@maps[[1]], size=0.5, spatial=FALSE)
}
\author{
Simon Moulds
}

