% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,ModelInput,ANY-method}
\alias{plot,ObsLulcMaps,ANY-method}
\alias{plot.ModelInput}
\alias{plot.ObsLulcMaps}
\title{Plot method for objects based on Raster* data}
\usage{
\method{plot}{ObsLulcMaps}(x, y, ...)

\method{plot}{ModelInput}(x, y, ...)

\S4method{plot}{ObsLulcMaps,ANY}(x, y, ...)

\S4method{plot}{ModelInput,ANY}(x, y, ...)
}
\arguments{
\item{x}{an ObsLulcMaps or *Model object}

\item{y}{not used}

\item{\dots}{additional arguments to
  \code{rasterVis::\link[rasterVis]{levelplot}}}
}
\value{
A trellis object.
}
\description{
Plot \code{lulcc} objects based on Raster* data
}
\examples{
## see lulccR-package examples
}
\seealso{
\code{rasterVis::\link[rasterVis]{levelplot}}
}

