% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Prediction}
\alias{Prediction}
\title{Create a Prediction object}
\usage{
Prediction(models, newdata, ...)
}
\arguments{
\item{models}{a PredModels object}

\item{newdata}{a data.frame containing new data}

\item{\dots}{additional arguments to \code{ROCR::\link[ROCR]{prediction}}}
}
\value{
A \code{Prediction} object.
}
\description{
This function creates a \code{ROCR::\link[ROCR]{prediction}} object for each
predictive model in a \code{PredModels} object. It should be used with
\code{\link{Performance}} to evaluate multiple models with exactly the
same criteria while keeping track of which model corresponds to which land use
category.
}
\examples{
## see lulccR-package examples
}
\references{
Sing, T., Sander, O., Beerenwinkel, N., Lengauer, T. (2005).
ROCR: visualizing classifier performance in R. Bioinformatics
21(20):3940-3941.
}
\seealso{
\code{link{Performance}}, \code{ROCR::\link[ROCR]{prediction}}
}

