% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{PredModels}
\alias{PredModels}
\alias{PredModels,list,ANY,ANY,numeric,character-method}
\alias{PredModels,list,ANY,numeric,character-method}
\alias{PredModels,list,ObsLulcMaps,ANY,ANY-method}
\title{Create a PredModels object}
\usage{
PredModels(models, obs, categories, labels, ...)

\S4method{PredModels}{list,ObsLulcMaps,ANY,ANY}(models, obs, categories, labels,
  ...)

\S4method{PredModels}{list,ANY,numeric,character}(models, obs, categories,
  labels, ...)
}
\arguments{
\item{models}{a list of predictive models}

\item{obs}{an ObsLulcMaps object}

\item{categories}{numeric vector of land use categories in observed maps. Only
required if \code{obs} is not provided}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}. Only required if \code{obs} is not provided}

\item{\dots}{additional arguments (none)}
}
\value{
An object of class \code{\link{PredModels}}.
}
\description{
Methods to create a \code{\link{PredModels-class}} object.
}
\details{
TODO
}
\examples{
## Plum Island Ecosystems
forest.glm <- glm(forest~1, family=binomial, data=train.data)
built.glm <- glm(built~ef_001+ef_002+ef_003, family=binomial, data=train.data)
other.glm <- glm(other~ef_001+ef_002, family=binomial, data=train.data)
}

