% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resample}
\alias{resample}
\alias{resample,ExpVarMaps,Raster-method}
\alias{resample,list,Raster-method}
\title{Resample maps in ExpVarMaps object or list}
\usage{
resample(x, y, ...)

\S4method{resample}{ExpVarMaps,Raster}(x, y, method = "ngb", ...)

\S4method{resample}{list,Raster}(x, y, method = "ngb", ...)
}
\arguments{
\item{x}{an ExpVarMaps object or list of Raster* maps to be resampled}

\item{y}{Raster* object with parameters that \code{x} should be resampled to}

\item{method}{method used to compute values for the new RasterLayer, should be
\code{"bilinear"} for bilinear interpolation, or \code{"ngb"} for nearest
neighbour}

\item{\dots}{additional arguments to \code{raster::\link[raster]{resample}}}
}
\value{
An ExpVarMaps object or list, depending on \code{x}.
}
\description{
A wrapper function for \code{raster::\link[raster]{resample}} to resample
raster objects in an ExpVarMaps object or list.
}
\examples{
\dontrun{

## create ExpVarMaps object
ef <- ExpVarMaps(x=pie, pattern="ef")

## resample to ensure maps have same characteristics as observed maps
ef <- resample(x=ef, y=pie$lu_pie_1985, method="ngb")
}
}
\seealso{
\code{\link{ExpVarMaps}}, \code{raster::\link[raster]{resample}}
}

