% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resample,PredictorMaps,Raster-method}
\alias{resample,PredictorMaps,Raster-method}
\alias{resample,list,Raster-method}
\title{Resample maps in PredictorMaps object or list}
\usage{
\S4method{resample}{PredictorMaps,Raster}(x, y, method = "ngb", ...)

\S4method{resample}{list,Raster}(x, y, method = "ngb", ...)
}
\arguments{
\item{x}{a PredictorMaps object or list of Raster* maps to be resampled}

\item{y}{Raster* object with parameters that \code{x} should be resampled to}

\item{method}{method used to compute values for the new RasterLayer, should be
\code{"bilinear"} for bilinear interpolation, or \code{"ngb"} for nearest
neighbour}

\item{\dots}{additional arguments to \code{raster::\link[raster]{resample}}}
}
\value{
A PredictorMaps object or list, depending on \code{x}.
}
\description{
A wrapper function for \code{raster::\link[raster]{resample}} to resample
raster objects in a \link{PredictorMaps-class} object or list.
}
\examples{
## create PredictorMaps object
pred <- predictorMaps(x=pie, pattern="pred")
pred <- PredictorMaps(maps=pred)

## resample to ensure maps have same characteristics as observed maps
pred <- resample(x=pred, y=pie$lu_pie_1985, method="ngb")
}
\seealso{
\code{\link{PredictorMaps}}, \code{raster::\link[raster]{resample}}
}

