% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roundSum}
\alias{roundSum}
\title{Round elements in matrix or data.frame rows}
\usage{
roundSum(x, ncell, ...)
}
\arguments{
\item{x}{matrix or data.frame}

\item{ncell}{numeric specifying the target sum for each row in \code{x}}

\item{\dots}{additional arguments (none)}
}
\value{
A matrix.
}
\description{
Round all numbers in a matrix or data.frame while ensuring that all rows sum
to the same value.
}
\details{
The main application of \code{roundSum} is to ensure that each row in the
demand matrix specifies exactly the number of cells to be allocated to each
land use category for the respective timestep. It may also be used to convert
the units of demand to number of cells, as required by \code{ModelInput}.
}
\examples{
## Sibuyan Island

## load demand scenario from data
dmd <- sibuyan$demand$demand1 * runif(1)
ncell <- length(which(!is.na(getValues(sibuyan$maps$lu_sib_1997))))

## recover demand
dmd <- roundSum(dmd, ncell=ncell)
}

