% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{subset,ExpVarMaps-method}
\alias{subset,ExpVarMaps-method}
\alias{subset,Performance-method}
\alias{subset,PredModels-method}
\alias{subset,Prediction-method}
\title{Subset}
\usage{
\S4method{subset}{ExpVarMaps}(x, subset, ...)

\S4method{subset}{PredModels}(x, subset, ...)

\S4method{subset}{Performance}(x, subset, ...)

\S4method{subset}{Prediction}(x, subset, ...)
}
\arguments{
\item{x}{an object of class \code{\link{ExpVarMaps}}, \code{\link{PredModels}}, #'   \code{Prediction} or \code{Performance}}

\item{subset}{integer or character indicating the objects to be extracted}

\item{...}{additional arguments (none)}
}
\description{
Extract a subset of objects from container classes such as \code{ExpVarMaps},
\code{PredModels}, \code{Prediction} and \code{Performance}.
}
\examples{
## Sibuyan Island

## load observed land use data
obs <- ObsLulcMaps(x=sibuyan$maps,
                    pattern="lu",
                    categories=c(1,2,3,4,5),
                    labels=c("Forest","Coconut","Grass","Rice","Other"),
                    t=c(0,14))

summary(obs)
obs <- subset(obs, subset=names(obs)[1])
summary(obs)

## load explanatory variables
ef <- ExpVarMaps(x=sibuyan$maps, pattern="ef")

summary(ef)
ef <- subset(ef, subset=1:10)
summary(ef)
}

