% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,ExpVarMaps-method}
\alias{summary,Model-method}
\alias{summary,ModelInput-method}
\alias{summary,NeighbMaps-method}
\alias{summary,ObsLulcMaps-method}
\alias{summary,PredModels-method}
\title{Summary}
\usage{
summary(object, ...)

\S4method{summary}{ObsLulcMaps}(object, ...)

\S4method{summary}{ExpVarMaps}(object, ...)

\S4method{summary}{NeighbMaps}(object, ...)

\S4method{summary}{PredModels}(object, ...)

\S4method{summary}{ModelInput}(object, ...)

\S4method{summary}{Model}(object, ...)
}
\arguments{
\item{object}{an object belonging to one of the classes in \code{lulcc}}

\item{...}{additional arguments (none)}
}
\value{
A matrix, data.frame or list
}
\description{
Summarise lulcc objects containing Raster* data or predictive models
}

