% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ThreeMapComparison}
\alias{ThreeMapComparison}
\title{Evaluate allocation performance with three maps}
\usage{
ThreeMapComparison(rt1, rt2, st2, categories, labels, factors, ...)
}
\arguments{
\item{rt1}{RasterLayer. Observed land use map at time 1}

\item{rt2}{RasterLayer. Observed land use map at time 2}

\item{st2}{RasterLayer. Simulated land use map at time 2}

\item{factors}{numeric vector of aggregation factors (equivalent to the 'fact'
argument to \code{raster::\link[raster]{aggregate}} representing the
resolutions at which model performance should be tested}

\item{...}{additional arguments to \code{raster::\link[raster]{aggregate}}}
}
\value{
a \code{ThreeMapComparison} object
}
\description{
An implementation of the method described by Pontius et al. (2011), which
compares a reference map at time 1, a reference map at time 2 and a simulated
map at time 2 to evaluate allocation performance at multiple resolutions while
taking into account persistence. The method quantifies disagreement within
coarse squares, disagreement between coarse squares, disagreement about the
quantity of land use change and agreement.
}
\author{
Simon Moulds
}
\references{
Pontius Jr, R.G., Peethambaram, S., Castella, J.C. (2011).
Comparison of three maps at multiple resolutions: a case study of land change
simulation in Cho Don District, Vietnam. Annals of the Association of American
Geographers 101(1): 45-62.
}
\seealso{
\code{\link{AgreementBudget}}, \code{\link{FigureOfMerit}}
}

